requirements.txt
Django>=3.0,<4.0
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-json-widget==1.0.1
django-redis>=4.12
django_pagedown>=2.2.0
djangorestframework==3.12.2
djangorestframework_simplejwt==4.6.0
gunicorn==20.0.4
pyjwt==2.1.0
psycopg2-binary>=2.8
